add PostgresQueryDriver class to bridge all db calls
this prepares us for a possible future where third parties supply custom
db drivers for non-postgres connections, while also helping to clarify
the Query class by separating out what is part of the query building
logic, vs what is part of the actual query execution context.

close https://rvohealth.atlassian.net/browse/PDTC-7967
