+++ package.json
@@ -1,7 +1,7 @@
 {
   "type": "module",
   "name": "@rvoh/dream",
-  "version": "0.44.7",
+  "version": "0.44.8",
   "description": "dream orm",
   "repository": {
     "type": "git",
