+++ src/dream/Query.ts
@@ -1,34 +1,10 @@
-import {
-  AliasedExpression,
-  DeleteQueryBuilder,
-  ExpressionBuilder,
-  ExpressionWrapper,
-  JoinBuilder,
-  ComparisonOperatorExpression as KyselyComparisonOperatorExpression,
-  SelectQueryBuilder,
-  SqlBool,
-  UpdateQueryBuilder,
-  Updateable,
-  sql,
-} from 'kysely'
-import pluralize from 'pluralize-esm'
-import ConnectedToDB from '../db/ConnectedToDB.js'
+import { DeleteQueryBuilder, SelectQueryBuilder, UpdateQueryBuilder } from 'kysely'
 import { SOFT_DELETE_SCOPE_NAME } from '../decorators/class/SoftDelete.js'
-import associationToGetterSetterProp from '../decorators/field/association/associationToGetterSetterProp.js'
 import DreamApp from '../dream-app/index.js'
 import Dream from '../Dream.js'
 import AssociationDeclaredWithoutAssociatedDreamClass from '../errors/associations/AssociationDeclaredWithoutAssociatedDreamClass.js'
-import CannotAssociateThroughPolymorphic from '../errors/associations/CannotAssociateThroughPolymorphic.js'
-import CannotJoinPolymorphicBelongsToError from '../errors/associations/CannotJoinPolymorphicBelongsToError.js'
-import JoinAttemptedOnMissingAssociation from '../errors/associations/JoinAttemptedOnMissingAssociation.js'
-import MissingRequiredAssociationAndClause from '../errors/associations/MissingRequiredAssociationAndClause.js'
-import MissingRequiredPassthroughForAssociationAndClause from '../errors/associations/MissingRequiredPassthroughForAssociationAndClause.js'
-import MissingThroughAssociation from '../errors/associations/MissingThroughAssociation.js'
-import MissingThroughAssociationSource from '../errors/associations/MissingThroughAssociationSource.js'
 import CannotCallUndestroyOnANonSoftDeleteModel from '../errors/CannotCallUndestroyOnANonSoftDeleteModel.js'
-import CannotNegateSimilarityClause from '../errors/CannotNegateSimilarityClause.js'
 import CannotPassAdditionalFieldsToPluckEachAfterCallback from '../errors/CannotPassAdditionalFieldsToPluckEachAfterCallback.js'
-import CannotPassUndefinedAsAValueToAWhereClause from '../errors/CannotPassUndefinedAsAValueToAWhereClause.js'
 import LeftJoinPreloadIncompatibleWithFindEach from '../errors/LeftJoinPreloadIncompatibleWithFindEach.js'
 import MissingRequiredCallbackFunctionToPluckEach from '../errors/MissingRequiredCallbackFunctionToPluckEach.js'
 import NoUpdateAllOnJoins from '../errors/NoUpdateAllOnJoins.js'
@@ -36,25 +12,11 @@ import NoUpdateOnAssociationQuery from '../errors/NoUpdateOnAssociationQuery.js'
 import CannotPaginateWithLimit from '../errors/pagination/CannotPaginateWithLimit.js'
 import CannotPaginateWithOffset from '../errors/pagination/CannotPaginateWithOffset.js'
 import RecordNotFound from '../errors/RecordNotFound.js'
-import UnexpectedUndefined from '../errors/UnexpectedUndefined.js'
-import CalendarDate from '../helpers/CalendarDate.js'
-import camelize from '../helpers/camelize.js'
 import cloneDeepSafe from '../helpers/cloneDeepSafe.js'
-import compact from '../helpers/compact.js'
-import { DateTime } from '../helpers/DateTime.js'
-import isEmpty from '../helpers/isEmpty.js'
 import isObject from '../helpers/isObject.js'
 import namespaceColumn from '../helpers/namespaceColumn.js'
-import normalizeUnicode from '../helpers/normalizeUnicode.js'
-import objectPathsToArrays from '../helpers/objectPathsToArrays.js'
 import protectAgainstPollutingAssignment from '../helpers/protectAgainstPollutingAssignment.js'
-import { Range } from '../helpers/range.js'
-import snakeify from '../helpers/snakeify.js'
-import uniq from '../helpers/uniq.js'
-import CurriedOpsStatement from '../ops/curried-ops-statement.js'
 import ops from '../ops/index.js'
-import OpsStatement from '../ops/ops-statement.js'
-import { BelongsToStatement } from '../types/associations/belongsTo.js'
 import { HasManyStatement } from '../types/associations/hasMany.js'
 import { HasOneStatement } from '../types/associations/hasOne.js'
 import {
@@ -64,21 +26,16 @@ import {
   OffsetStatement,
   OrderQueryStatement,
   PassthroughOnClause,
-  SelfOnStatement,
   WhereStatement,
   WhereStatementForJoinedAssociation,
 } from '../types/associations/shared.js'
 import { DbConnectionType } from '../types/db.js'
 import {
-  AliasToDreamIdMap,
   AllDefaultScopeNames,
-  AssociationNameToAssociationDataAndDreamClassMap,
-  AssociationNameToAssociationMap,
-  AssociationNameToDreamClassMap,
   DefaultScopeName,
   DreamColumnNames,
+  DreamConstructorType,
   DreamTableSchema,
-  JoinAndStatements,
   OrderDir,
   PassthroughColumnNames,
   PluckEachArgs,
@@ -94,11 +51,9 @@ import {
   DefaultQueryTypeOptions,
   ExtendQueryType,
   FindEachOpts,
-  JoinTypes,
   NamespacedOrBaseModelColumnTypes,
   PaginatedDreamQueryOptions,
   PaginatedDreamQueryResult,
-  PreloadedDreamsAndWhatTheyPointTo,
   QueryToKyselyDBType,
   QueryToKyselyTableNamesType,
 } from '../types/query.js'
@@ -110,20 +65,14 @@ import {
   VariadicLeftJoinLoadArgs,
   VariadicLoadArgs,
 } from '../types/variadic.js'
-import { DreamConst } from './constants.js'
 import DreamTransaction from './DreamTransaction.js'
 import computedPaginatePage from './internal/computedPaginatePage.js'
-import executeDatabaseQuery from './internal/executeDatabaseQuery.js'
-import extractAssociationMetadataFromAssociationName from './internal/extractAssociationMetadataFromAssociationName.js'
-import orderByDirection from './internal/orderByDirection.js'
-import shouldBypassDefaultScope from './internal/shouldBypassDefaultScope.js'
-import SimilarityBuilder from './internal/similarity/SimilarityBuilder.js'
-import sqlResultToDreamInstance from './internal/sqlResultToDreamInstance.js'
+import PostgresQueryDriver from './QueryDriver/Postgres.js'
 
 export default class Query<
   DreamInstance extends Dream,
   QueryTypeOpts extends Readonly<QueryTypeOptions> = DefaultQueryTypeOptions<DreamInstance>,
-> extends ConnectedToDB<DreamInstance> {
+> {
   /**
    * @internal
    *
@@ -148,7 +97,7 @@ export default class Query<
    * stores the dream transaction applied to the
    * current Query instance
    */
-  public override dreamTransaction: DreamTransaction<Dream> | null = null
+  public dreamTransaction: DreamTransaction<Dream> | null = null
 
   /**
    * @internal
@@ -352,11 +301,33 @@ export default class Query<
    */
   private baseSelectQuery: Query<any, any> | null
 
+  /*
+   * Store the original opts, so we can hand them off succinctly to the
+   * query executor. TODO: Should we do this?
+   * */
+  private originalOpts: QueryOpts<DreamInstance, DreamColumnNames<DreamInstance>>
+
+  protected readonly dreamClass: DreamConstructorType<DreamInstance>
+  public readonly dreamInstance: DreamInstance
+  protected connectionOverride: DbConnectionType | undefined
+
+  /**
+   * @internal
+   *
+   * stores the Dream models joined in this Query instance
+   */
+  protected readonly innerJoinDreamClasses: readonly (typeof Dream)[] = Object.freeze([])
+
   constructor(
     dreamInstance: DreamInstance,
     opts: QueryOpts<DreamInstance, DreamColumnNames<DreamInstance>> = {}
   ) {
-    super(dreamInstance, opts)
+    this.dreamInstance = dreamInstance
+    this.dreamClass = dreamInstance.constructor as DreamConstructorType<DreamInstance>
+    this.dreamTransaction = opts.transaction || null
+    this.connectionOverride = opts.connection
+    this.innerJoinDreamClasses = Object.freeze(opts.innerJoinDreamClasses || [])
+
     this.passthroughOnStatement = Object.freeze(opts.passthroughOnStatement || {})
     this.whereStatements = Object.freeze(opts.where || [])
     this.whereNotStatements = Object.freeze(opts.whereNot || [])
@@ -381,6 +352,7 @@ export default class Query<
     this.distinctColumn = opts.distinctColumn || null
     this.connectionOverride = opts.connection
     this.shouldReallyDestroy = opts.shouldReallyDestroy || false
+    this.originalOpts = Object.freeze(opts)
   }
 
   /**
@@ -391,37 +363,6 @@ export default class Query<
    */
   public get isDreamQuery() {
     return true
-  }
-
-  /**
-   * @internal
-   *
-   * Used for applying preload and load statements
-   *
-   * @returns An associated Query
-   */
-  private dreamClassQueryWithScopeBypasses<T extends typeof Dream>(
-    dreamClass: T,
-    {
-      bypassAllDefaultScopesExceptOnAssociations = false,
-      defaultScopesToBypassExceptOnAssociations = [],
-    }: {
-      bypassAllDefaultScopesExceptOnAssociations?: boolean | undefined
-      defaultScopesToBypassExceptOnAssociations?: AllDefaultScopeNames<DreamInstance>[] | undefined
-    } = {}
-  ): Query<InstanceType<T>, DefaultQueryTypeOptions<DreamInstance>> {
-    const associationQuery = dreamClass.query().clone({
-      passthroughOnStatement: this.passthroughOnStatement,
-      bypassAllDefaultScopes: this.bypassAllDefaultScopes,
-      bypassAllDefaultScopesExceptOnAssociations,
-      defaultScopesToBypass: this.defaultScopesToBypass,
-      defaultScopesToBypassExceptOnAssociations,
-    })
-
-    return (this.dreamTransaction ? associationQuery.txn(this.dreamTransaction) : associationQuery) as Query<
-      InstanceType<T>,
-      DefaultQueryTypeOptions<DreamInstance>
-    >
   }
 
   /**
@@ -925,128 +866,6 @@ export default class Query<
   /**
    * @internal
    *
-   *
-   */
-  private associationNamesToDreamClassesMap(
-    associationNames: string[],
-    associationsToDreamClassesMap: AssociationNameToDreamClassMap = {}
-  ): AssociationNameToDreamClassMap {
-    const namesToAssociationsAndDreamClasses =
-      this.associationNamesToAssociationDataAndDreamClassesMap(associationNames)
-    return Object.keys(namesToAssociationsAndDreamClasses).reduce((remap, associationName) => {
-      const associationAndDreamClass = namesToAssociationsAndDreamClasses[associationName]
-      if (associationAndDreamClass === undefined) throw new UnexpectedUndefined()
-
-      remap[associationName] = associationAndDreamClass.dreamClass
-      return remap
-    }, associationsToDreamClassesMap)
-  }
-
-  /**
-   * @internal
-   *
-   *
-   */
-  private associationNamesToAssociationsMap(
-    associationNames: string[],
-    associationsToAssociations: AssociationNameToAssociationMap = {}
-  ): AssociationNameToAssociationMap {
-    const namesToAssociationsAndDreamClasses =
-      this.associationNamesToAssociationDataAndDreamClassesMap(associationNames)
-    return Object.keys(namesToAssociationsAndDreamClasses).reduce((remap, associationName) => {
-      const associationAndDreamClass = namesToAssociationsAndDreamClasses[associationName]
-      if (associationAndDreamClass === undefined) throw new UnexpectedUndefined()
-
-      remap[associationName] = associationAndDreamClass.association
-      return remap
-    }, associationsToAssociations)
-  }
-
-  /**
-   * @internal
-   */
-  private associationNamesToAssociationDataAndDreamClassesMap(
-    associationNames: string[]
-  ): AssociationNameToAssociationDataAndDreamClassMap {
-    const associationsToDreamClassesMap: AssociationNameToAssociationDataAndDreamClassMap = {}
-
-    associationNames.reduce((dreamClass: typeof Dream, associationName: string) => {
-      const { name, alias } = extractAssociationMetadataFromAssociationName(associationName)
-      const association = dreamClass['getAssociationMetadata'](name)
-      if (association === undefined) throw new UnexpectedUndefined()
-
-      const through = (association as any).through
-
-      if (through) {
-        const { throughAssociation, throughAssociationDreamClass } = this.throughAssociationDetails(
-          dreamClass,
-          through
-        )
-        associationsToDreamClassesMap[through] = {
-          association: throughAssociation,
-          dreamClass: throughAssociationDreamClass,
-        }
-      }
-
-      const nextDreamClass = association.modelCB() as typeof Dream
-      if (alias === undefined) throw new UnexpectedUndefined()
-      associationsToDreamClassesMap[alias] = { association, dreamClass: nextDreamClass }
-      return nextDreamClass
-    }, this.dreamClass)
-
-    return associationsToDreamClassesMap
-  }
-
-  /**
-   * @internal
-   */
-  private throughAssociationDetails(
-    dreamClass: typeof Dream,
-    through: string
-  ): {
-    throughAssociation: AssociationStatement
-    throughAssociationDreamClass: typeof Dream
-  } {
-    const throughAssociation = dreamClass['getAssociationMetadata'](through)
-    if (throughAssociation === undefined) throw new UnexpectedUndefined()
-
-    const throughAssociationDreamClass = throughAssociation.modelCB() as typeof Dream
-    return { throughAssociation, throughAssociationDreamClass }
-  }
-
-  /**
-   * @internal
-   *
-   *
-   */
-  private joinStatementsToDreamClassesMap(joinStatements: RelaxedJoinStatement) {
-    const associationsToDreamClassesMap: AssociationNameToDreamClassMap = {}
-
-    objectPathsToArrays(joinStatements).forEach(associationChain =>
-      this.associationNamesToDreamClassesMap(associationChain, associationsToDreamClassesMap)
-    )
-
-    return associationsToDreamClassesMap
-  }
-
-  /**
-   * @internal
-   *
-   *
-   */
-  private joinStatementsToAssociationsMap(joinStatements: RelaxedJoinStatement) {
-    const associationsToAssociationsMap: AssociationNameToAssociationMap = {}
-
-    objectPathsToArrays(joinStatements).forEach(associationChain =>
-      this.associationNamesToAssociationsMap(associationChain, associationsToAssociationsMap)
-    )
-
-    return associationsToAssociationsMap
-  }
-
-  /**
-   * @internal
-   *
    * Adds Dream class to temporary innerJoinDreamClasses array based on association name
    *
    */
@@ -1337,8 +1156,7 @@ export default class Query<
     this: Query<DreamInstance, QueryTypeOpts>,
     selection: SimpleFieldType | PluckThroughFieldType
   ) {
-    const query = this.buildSelect({ bypassSelectAll: true, bypassOrder: true })
-    return query.select(this.namespaceColumn(selection as any))
+    return this.dbDriverInstance().nestedSelect(selection)
   }
 
   /**
@@ -1483,8 +1301,7 @@ export default class Query<
    *
    */
   public sql() {
-    const kyselyQuery = this.buildSelect()
-    return kyselyQuery.compile()
+    return this.dbDriverInstance().sql()
   }
 
   /**
@@ -1509,25 +1326,8 @@ export default class Query<
         : QueryType extends 'update'
           ? UpdateQueryBuilder<DbType, TableNames & keyof DbType, TableNames & keyof DbType, unknown>
           : never,
-  >(type: QueryType) {
-    switch (type) {
-      case 'select':
-        return this.buildSelect() as ToKyselyReturnType
-
-      case 'delete':
-        return this.buildDelete() as ToKyselyReturnType
-
-      case 'update':
-        return this.buildUpdate({}) as ToKyselyReturnType
-
-      // TODO: in the future, we should support insert type, but don't yet, since inserts are done outside
-      // the query class for some reason.
-      default: {
-        // protection so that if a new QueryType is ever added, this will throw a type error at build time
-        const _never: never = type
-        throw new Error(`Unhandled QueryType: ${_never as string}`)
-      }
-    }
+  >(type: QueryType): ToKyselyReturnType {
+    return this.dbDriverInstance().toKysely(type)
   }
 
   /**
@@ -1557,22 +1357,7 @@ export default class Query<
    * @returns The number of records in the database
    */
   public async count() {
-    // eslint-disable-next-line @typescript-eslint/unbound-method
-    const { count } = this.dbFor('select').fn
-    const distinctColumn = this.distinctColumn
-    const query = this.clone({ distinctColumn: null })
-
-    let kyselyQuery = query.buildSelect({ bypassSelectAll: true, bypassOrder: true })
-
-    const countClause = distinctColumn
-      ? count(sql`DISTINCT ${distinctColumn}`)
-      : count(query.namespaceColumn(query.dreamInstance.primaryKey))
-
-    kyselyQuery = kyselyQuery.select(countClause.as('tablecount'))
-
-    const data = await executeDatabaseQuery(kyselyQuery, 'executeTakeFirstOrThrow')
-
-    return parseInt(data.tablecount.toString())
+    return await this.dbDriverInstance().count()
   }
 
   /**
@@ -1633,15 +1418,7 @@ export default class Query<
     >,
     ReturnType extends NamespacedOrBaseModelColumnTypes<[ColumnName], Q, DreamInstance>[0],
   >(columnName: ColumnName): Promise<ReturnType> {
-    // eslint-disable-next-line @typescript-eslint/unbound-method
-    const { max } = this.dbFor('select').fn
-    let kyselyQuery = this.buildSelect({ bypassSelectAll: true, bypassOrder: true })
-
-    kyselyQuery = kyselyQuery.select(max(columnName as any) as any)
-
-    const data = await executeDatabaseQuery(kyselyQuery, 'executeTakeFirstOrThrow')
-
-    return data.max
+    return await this.dbDriverInstance().max(columnName)
   }
 
   /**
@@ -1667,283 +1444,7 @@ export default class Query<
     >,
     ReturnType extends NamespacedOrBaseModelColumnTypes<[ColumnName], Q, DreamInstance>[0],
   >(columnName: ColumnName): Promise<ReturnType> {
-    // eslint-disable-next-line @typescript-eslint/unbound-method
-    const { min } = this.dbFor('select').fn
-    let kyselyQuery = this.buildSelect({ bypassSelectAll: true, bypassOrder: true })
-
-    kyselyQuery = kyselyQuery.select(min(columnName as any) as any)
-    const data = await executeDatabaseQuery(kyselyQuery, 'executeTakeFirstOrThrow')
-
-    return data.min
-  }
-
-  /**
-   * @internal
-   *
-   * Runs the query and extracts plucked values
-   *
-   * @returns An array of plucked values
-   */
-  private async executePluck(...fields: DreamColumnNames<DreamInstance>[]): Promise<any[]> {
-    let kyselyQuery = this.removeAllDefaultScopesExceptOnAssociations().buildSelect({ bypassSelectAll: true })
-    const aliases: string[] = []
-
-    fields.forEach((field: string) => {
-      // field will already be namespaced in a join situation, but when the field to pluck is on the
-      // base model, it will be underscored (to match the table name), but when the selected column
-      // comes back from Kysely camelCased
-      aliases.push(field.includes('_') ? camelize(field) : field)
-      //  namespace the selection so that when plucking the same column name from
-      // multpile tables, they don't get saved as the same name (e.g. select results with two `id` columns,
-      // which the pg package then returns in an object with a single `id` key)
-      kyselyQuery = kyselyQuery.select(`${this.namespaceColumn(field)} as ${field}` as any)
-    })
-
-    return (await executeDatabaseQuery(kyselyQuery, 'execute')).map(singleResult =>
-      aliases.map(alias => singleResult[alias])
-    )
-  }
-
-  /**
-   * @internal
-   *
-   */
-  private async executeJoinLoad(
-    options: {
-      columns?: DreamColumnNames<DreamInstance>[]
-    } = {}
-  ): Promise<DreamInstance[]> {
-    const query = (this as any).limit(null).offset(null)
-
-    let kyselyQuery = query.buildSelect({ bypassSelectAll: true })
-
-    const aliasToDreamClassesMap = {
-      [this.baseSqlAlias]: this.dreamClass,
-      ...this.joinStatementsToDreamClassesMap(this.leftJoinStatements),
-    }
-
-    const associationAliasToColumnAliasMap: Record<string, Record<string, string>> = {}
-    const aliasToAssociationsMap = this.joinStatementsToAssociationsMap(this.leftJoinStatements)
-
-    const aliases = Object.keys(aliasToDreamClassesMap)
-
-    let nextColumnAliasCounter = 0
-
-    aliases.forEach((aliasOrExpression: string) => {
-      const alias = extractAssociationMetadataFromAssociationName(aliasOrExpression).alias
-      if (alias === undefined) throw new UnexpectedUndefined()
-
-      associationAliasToColumnAliasMap[alias] ||= {}
-      const aliasedDreamClass = aliasToDreamClassesMap[alias]
-      if (aliasedDreamClass === undefined) throw new UnexpectedUndefined()
-      const association = aliasToAssociationsMap[alias]
-
-      const columns =
-        alias === this.baseSqlAlias
-          ? options.columns
-            ? this.columnsWithRequiredLoadColumns(options.columns)
-            : this.dreamClass.columns()
-          : aliasedDreamClass.columns()
-
-      columns.forEach((column: string) => {
-        const columnAlias = `dr${nextColumnAliasCounter++}`
-        kyselyQuery = kyselyQuery.select(`${this.namespaceColumn(column, alias)} as ${columnAlias}`)
-        const columnAliasMap = associationAliasToColumnAliasMap[alias]
-        if (columnAliasMap === undefined) throw new UnexpectedUndefined()
-
-        columnAliasMap[column] = columnAlias
-      })
-
-      if (association?.type === 'HasOne' || association?.type === 'HasMany') {
-        const setupPreloadData = (dbColumnName: string) => {
-          const columnAlias = `dr${nextColumnAliasCounter++}`
-          const columnAliasMap = associationAliasToColumnAliasMap[association.through!]
-          if (columnAliasMap === undefined) throw new UnexpectedUndefined()
-
-          columnAliasMap[dbColumnName] = columnAlias
-          kyselyQuery = kyselyQuery.select(
-            `${this.namespaceColumn(dbColumnName, association.through)} as ${columnAlias}`
-          )
-        }
-
-        if (association.through && association.preloadThroughColumns) {
-          if (isObject(association.preloadThroughColumns)) {
-            const preloadMap = association.preloadThroughColumns as Record<string, string>
-            Object.keys(preloadMap).forEach(columnName => setupPreloadData(columnName))
-          } else {
-            const preloadArray = association.preloadThroughColumns as string[]
-            preloadArray.forEach(columnName => setupPreloadData(columnName))
-          }
-        }
-      }
-    })
-
-    const queryResults = await executeDatabaseQuery(kyselyQuery, 'execute')
-
-    const aliasToDreamIdMap = queryResults.reduce(
-      (aliasToDreamIdMap: AliasToDreamIdMap, singleSqlResult: any) => {
-        this.fleshOutJoinLoadExecutionResults({
-          currentAlias: this.baseSqlAlias,
-          singleSqlResult,
-          aliasToDreamIdMap,
-          associationAliasToColumnAliasMap,
-          aliasToAssociationsMap,
-          aliasToDreamClassesMap,
-          leftJoinStatements: this.leftJoinStatements,
-        })
-
-        return aliasToDreamIdMap
-      },
-      {} as AliasToDreamIdMap
-    )
-
-    const baseModelIdToDreamMap = aliasToDreamIdMap[this.baseSqlAlias] || new Map()
-    return compact(Array.from(baseModelIdToDreamMap.values()) as DreamInstance[])
-  }
-
-  private fleshOutJoinLoadExecutionResults({
-    currentAlias,
-    singleSqlResult,
-    aliasToDreamIdMap,
-    associationAliasToColumnAliasMap,
-    aliasToAssociationsMap,
-    aliasToDreamClassesMap,
-    leftJoinStatements,
-  }: {
-    currentAlias: string
-    singleSqlResult: any
-    aliasToDreamIdMap: AliasToDreamIdMap
-    associationAliasToColumnAliasMap: Record<string, Record<string, string>>
-    aliasToAssociationsMap: AssociationNameToAssociationMap
-    aliasToDreamClassesMap: AssociationNameToDreamClassMap
-    leftJoinStatements: RelaxedJoinStatement
-  }) {
-    const dreamClass = aliasToDreamClassesMap[currentAlias]
-    if (dreamClass === undefined) throw new UnexpectedUndefined()
-
-    const columnToColumnAliasMap = associationAliasToColumnAliasMap[currentAlias]
-    if (columnToColumnAliasMap === undefined) throw new UnexpectedUndefined()
-
-    const primaryKeyName = dreamClass.primaryKey
-    if (primaryKeyName === undefined) throw new UnexpectedUndefined()
-
-    const columnAlias = columnToColumnAliasMap[primaryKeyName]
-    if (columnAlias === undefined) throw new UnexpectedUndefined()
-
-    const primaryKeyValue = singleSqlResult[columnAlias]
-    if (!primaryKeyValue) return null
-
-    aliasToDreamIdMap[currentAlias] ||= new Map()
-
-    if (!aliasToDreamIdMap[currentAlias].get(primaryKeyValue)) {
-      const columnValueMap = Object.keys(columnToColumnAliasMap).reduce(
-        (columnNameValueMap, columnName) => {
-          const columnAlias = columnToColumnAliasMap[columnName]
-          if (columnAlias === undefined) throw new UnexpectedUndefined()
-          columnNameValueMap[columnName] = singleSqlResult[columnAlias]
-          return columnNameValueMap
-        },
-        {} as Record<string, any>
-      )
-      const dream = sqlResultToDreamInstance(dreamClass, columnValueMap)
-
-      const association = aliasToAssociationsMap[currentAlias] as
-        | HasOneStatement<any, any, any, any>
-        | HasManyStatement<any, any, any, any>
-      if (association && association.through && association.preloadThroughColumns) {
-        const throughAssociationColumnToColumnAliasMap = associationAliasToColumnAliasMap[association.through]
-        if (throughAssociationColumnToColumnAliasMap === undefined) throw new UnexpectedUndefined()
-
-        this.hydratePreloadedThroughColumns({
-          association,
-          columnToColumnAliasMap: throughAssociationColumnToColumnAliasMap,
-          dream,
-          singleSqlResult,
-        })
-      }
-
-      aliasToDreamIdMap[protectAgainstPollutingAssignment(currentAlias)]?.set(primaryKeyValue, dream)
-    }
-
-    const dream = aliasToDreamIdMap[currentAlias].get(primaryKeyValue)
-
-    Object.keys(leftJoinStatements).forEach(nextAlias => {
-      const { name: associationName, alias } = extractAssociationMetadataFromAssociationName(nextAlias)
-
-      const association = dreamClass['getAssociationMetadata'](associationName)
-      if (association === undefined) throw new UnexpectedUndefined()
-
-      const associatedDream = this.fleshOutJoinLoadExecutionResults({
-        currentAlias: alias,
-        singleSqlResult,
-        aliasToDreamIdMap,
-        associationAliasToColumnAliasMap,
-        aliasToAssociationsMap,
-        aliasToDreamClassesMap,
-        leftJoinStatements: leftJoinStatements[nextAlias] as RelaxedJoinStatement,
-      })
-      const hasMany = association.type === 'HasMany'
-
-      // initialize by trying to access the association, which throws an exception if not yet initialized
-      try {
-        ;(dream as any)[association.as]
-      } catch {
-        if (hasMany) (dream as any)[association.as] = []
-        else (dream as any)[associationToGetterSetterProp(association)] = null
-      }
-
-      if (!associatedDream) return
-
-      if (hasMany) {
-        if (!(dream as any)[association.as].includes(associatedDream))
-          (dream as any)[association.as].push(associatedDream)
-      } else (dream as any)[associationToGetterSetterProp(association)] = associatedDream
-    })
-
-    return dream
-  }
-
-  private hydratePreloadedThroughColumns({
-    association,
-    columnToColumnAliasMap,
-    dream,
-    singleSqlResult,
-  }: {
-    association: HasOneStatement<any, any, any, any> | HasManyStatement<any, any, any, any>
-    columnToColumnAliasMap: Record<string, string>
-    dream: Dream
-    singleSqlResult: any
-  }) {
-    if (!association.through) return
-    if (!(dream as any).preloadedThroughColumns) return
-
-    let columnNames: string[] = []
-    const columnNameToPreloadedThroughColumnNameMap: Record<string, string> = {}
-
-    if (isObject(association.preloadThroughColumns)) {
-      const preloadMap = association.preloadThroughColumns as Record<string, string>
-      columnNames = Object.keys(preloadMap).map(columnName => {
-        columnNameToPreloadedThroughColumnNameMap[columnName] = preloadMap[columnName]!
-        return columnName
-      })
-    } else if (Array.isArray(association.preloadThroughColumns)) {
-      columnNames = association.preloadThroughColumns.map(columnName => {
-        columnNameToPreloadedThroughColumnNameMap[columnName] = columnName
-        return columnName
-      })
-    }
-
-    columnNames.forEach(columnName => {
-      const preloadedThroughColumnName = columnNameToPreloadedThroughColumnNameMap[columnName]
-      if (preloadedThroughColumnName === undefined) throw new UnexpectedUndefined()
-
-      const columnAlias = columnToColumnAliasMap[columnName]
-      if (columnAlias === undefined) {
-        throw new UnexpectedUndefined()
-      }
-
-      ;(dream as any).preloadedThroughColumns[preloadedThroughColumnName] = singleSqlResult[columnAlias]
-    })
+    return await this.dbDriverInstance().min(columnName)
   }
 
   /**
@@ -1978,7 +1479,7 @@ export default class Query<
       ? NamespacedOrBaseModelColumnTypes<ColumnNames, Q, DreamInstance>[0][]
       : NamespacedOrBaseModelColumnTypes<ColumnNames, Q, DreamInstance>[],
   >(this: Q, ...columnNames: ColumnNames): Promise<ReturnValue> {
-    const vals = await this.executePluck(...(columnNames as any[]))
+    const vals = await this.dbDriverInstance().pluck(...(columnNames as any[]))
 
     return (columnNames.length > 1 ? vals : vals.flat()) as ReturnValue
   }
@@ -2035,11 +1536,12 @@ export default class Query<
         ? onlyColumns
         : [this.namespacedPrimaryKey, ...onlyColumns]
 
-      records = await this.offset(offset as any)
+      const query = this.offset(offset as any)
         .order(null)
         .order(this.namespacedPrimaryKey as any)
         .limit(batchSize as any)
-        .executePluck(...columnsIncludingPrimaryKey)
+
+      records = await this.dbDriverInstance(query).pluck(...columnsIncludingPrimaryKey)
 
       // In order to batch, we need to order by primary key, so the primary key must be plucked.
       // If the developer did not include the primary key in the columns to pluck, then we prepended it above
@@ -2072,14 +1574,16 @@ export default class Query<
       columns?: DreamColumnNames<DreamInstance>[]
     } = {}
   ) {
-    if (this.joinLoadActivated) return await this.executeJoinLoad(options)
+    return await this.dbDriverInstance().takeAll(options)
+  }
 
-    const kyselyQuery = this.buildSelect(options)
-    const results = await executeDatabaseQuery(kyselyQuery, 'execute')
-    const theAll = results.map(r => sqlResultToDreamInstance(this.dreamClass, r)) as DreamInstance[]
-    await this.applyPreload(this.preloadStatements as any, this.preloadOnStatements as any, theAll)
+  public static dbDriverClass<T extends Dream>() {
+    return PostgresQueryDriver<T>
+  }
 
-    return theAll
+  public dbDriverInstance(query: Query<DreamInstance, any> = this) {
+    const driverClass = Query.dbDriverClass<DreamInstance>()
+    return new driverClass(query)
   }
 
   /**
@@ -2197,7 +1701,8 @@ export default class Query<
     const query = this.orderStatements.length
       ? this
       : this.order({ [this.namespacedPrimaryKey as any]: 'asc' } as any)
-    return await query.takeOne()
+    const dbDriverClass = Query.dbDriverClass<DreamInstance>()
+    return await new dbDriverClass(query).takeOne()
   }
 
   /**
@@ -2238,7 +1743,8 @@ export default class Query<
       ? this.invertOrder()
       : this.order({ [this.namespacedPrimaryKey]: 'desc' } as any)
 
-    return await query.takeOne()
+    const dbDriverClass = Query.dbDriverClass<DreamInstance>()
+    return await new dbDriverClass(query).takeOne()
   }
 
   /**
@@ -2414,8 +1920,7 @@ export default class Query<
    * @returns The number of records that were removed
    */
   public async delete(): Promise<number> {
-    const deletionResult = await executeDatabaseQuery(this.buildDelete(), 'executeTakeFirst')
-    return Number(deletionResult?.numDeletedRows || 0)
+    return await this.dbDriverInstance().delete()
   }
 
   /**
@@ -2455,1565 +1960,7 @@ export default class Query<
   }
 
   private async updateWithoutCallingModelHooks(attributes: DreamTableSchema<DreamInstance>) {
-    const kyselyQuery = this.buildUpdate(attributes)
-    const res = await executeDatabaseQuery(kyselyQuery, 'execute')
-    const resultData = Array.from(res.entries())?.[0]?.[1]
-
-    return Number(resultData?.numUpdatedRows || 0)
-  }
-
-  /**
-   * @internal
-   *
-   * Used for applying first and last queries
-   *
-   * @returns A dream instance or null
-   */
-  private async takeOne() {
-    if (this.joinLoadActivated) {
-      let query: Query<DreamInstance, QueryTypeOpts>
-
-      if (
-        this.whereStatements.find(
-          whereStatement =>
-            (whereStatement as any)[this.dreamClass.primaryKey] ||
-            (whereStatement as any)[this.namespacedPrimaryKey]
-        )
-      ) {
-        // the query already includes a primary key where statement
-        query = this
-      } else {
-        // otherwise find the primary key and apply it to the query
-        const primaryKeyValue = (await (this as any).limit(1).pluck(this.namespacedPrimaryKey as any))[0]
-        if (primaryKeyValue === undefined) return null
-        query = this.where({ [this.namespacedPrimaryKey]: primaryKeyValue } as any)
-      }
-
-      return (await query.executeJoinLoad())[0] || null
-    }
-
-    const kyselyQuery = (this as any).limit(1).buildSelect()
-    const results = await executeDatabaseQuery(kyselyQuery, 'executeTakeFirst')
-
-    if (results) {
-      const theFirst = sqlResultToDreamInstance(this.dreamClass, results) as DreamInstance
-
-      if (theFirst)
-        await this.applyPreload(this.preloadStatements as any, this.preloadOnStatements as any, [theFirst])
-
-      return theFirst
-    } else return null
-  }
-
-  /**
-   * @internal
-   *
-   * Used to hydrate dreams with the provided associations
-   */
-  private hydrateAssociation(
-    dreams: Dream[],
-    association: AssociationStatement,
-    preloadedDreamsAndWhatTheyPointTo: PreloadedDreamsAndWhatTheyPointTo[]
-  ) {
-    switch (association.type) {
-      case 'HasMany':
-        dreams.forEach((dream: any) => {
-          dream[association.as] = []
-        })
-        break
-      default:
-        dreams.forEach((dream: any) => {
-          dream[associationToGetterSetterProp(association)] = null
-        })
-    }
-
-    // dreams is a Rating
-    // Rating belongs to: rateables (Posts / Compositions)
-    // loadedAssociations is an array of Posts and Compositions
-    // if rating.rateable_id === loadedAssociation.primaryKeyvalue
-    //  rating.rateable = loadedAssociation
-
-    preloadedDreamsAndWhatTheyPointTo.forEach(preloadedDreamAndWhatItPointsTo => {
-      dreams
-        .filter(dream => dream.primaryKeyValue === preloadedDreamAndWhatItPointsTo.pointsToPrimaryKey)
-        .forEach((dream: any) => {
-          if (association.type === 'HasMany') {
-            dream[association.as].push(preloadedDreamAndWhatItPointsTo.dream)
-          } else {
-            // in a HasOne context, order clauses will be applied in advance,
-            // prior to hydration. Considering, we only want to set the first
-            // result and ignore other results, so we will use ||= to set.
-            dream[association.as] ||= preloadedDreamAndWhatItPointsTo.dream
-          }
-        })
-    })
-
-    if (association.type === 'HasMany') {
-      dreams.forEach((dream: any) => Object.freeze(dream[association.as]))
-    }
-  }
-
-  /**
-   * @internal
-   *
-   * Used to bridge through associations
-   */
-  private followThroughAssociation(
-    dreamClass: typeof Dream,
-    association: HasOneStatement<any, any, any, any> | HasManyStatement<any, any, any, any>
-  ) {
-    const throughAssociation =
-      association.through && dreamClass['getAssociationMetadata'](association.through)
-    if (!throughAssociation)
-      throw new MissingThroughAssociation({
-        dreamClass,
-        association,
-      })
-
-    const throughClass = throughAssociation.modelCB() as typeof Dream
-    if (Array.isArray(throughClass))
-      throw new CannotAssociateThroughPolymorphic({
-        dreamClass,
-        association,
-      })
-
-    const newAssociation = getSourceAssociation(throughClass, association.source)
-    if (!newAssociation)
-      throw new MissingThroughAssociationSource({
-        dreamClass,
-        throughClass,
-        association,
-      })
-
-    return { throughAssociation, throughClass, newAssociation }
-  }
-
-  /**
-   * @internal
-   *
-   * Polymorphic BelongsTo. Since polymorphic associations may point to multiple tables,
-   * preload by loading each target class separately.
-   *
-   * Used to preload polymorphic belongs to associations
-   */
-  private async preloadPolymorphicBelongsTo(
-    this: Query<DreamInstance, QueryTypeOpts>,
-    association: BelongsToStatement<any, any, any, string>,
-    dreams: Dream[]
-  ) {
-    if (!association.polymorphic)
-      throw new Error(
-        `Association ${association.as} points to an array of models but is not designated polymorphic`
-      )
-    if (association.type !== 'BelongsTo')
-      throw new Error(
-        `Polymorphic association ${association.as} points to an array of models but is ${association.type as string}. Only BelongsTo associations may point to an array of models.`
-      )
-
-    const associatedDreams: Dream[] = []
-
-    for (const associatedModel of association.modelCB() as (typeof Dream)[]) {
-      await this.preloadPolymorphicAssociationModel(dreams, association, associatedModel, associatedDreams)
-    }
-
-    return associatedDreams
-  }
-
-  private async preloadPolymorphicAssociationModel(
-    dreams: Dream[],
-    association: BelongsToStatement<any, any, any, string>,
-    associatedDreamClass: typeof Dream,
-    associatedDreams: Dream[]
-  ) {
-    const relevantAssociatedModels = dreams.filter((dream: any) => {
-      const field = association.foreignKeyTypeField()
-      return dream[field] === associatedDreamClass['stiBaseClassOrOwnClassName'] || dream[field] === null
-    })
-
-    if (relevantAssociatedModels.length) {
-      dreams.forEach((dream: any) => {
-        dream[associationToGetterSetterProp(association)] = null
-      })
-
-      // Load all models of type associated that are associated with any of the already loaded Dream models
-      const loadedAssociations = await this.dreamClassQueryWithScopeBypasses(associatedDreamClass, {
-        // The association may remove specific default scopes that would otherwise preclude
-        // certain instances of the associated class from being found.
-        defaultScopesToBypassExceptOnAssociations: association.withoutDefaultScopes,
-      })
-        .where({
-          [associatedDreamClass.primaryKey]: relevantAssociatedModels.map(
-            (dream: any) => dream[association.foreignKey()]
-          ),
-        })
-        .all()
-
-      loadedAssociations.forEach((loadedAssociation: Dream) => associatedDreams.push(loadedAssociation))
-
-      //////////////////////////////////////////////////////////////////////////////////////////////
-      // Associate each loaded association with each dream based on primary key and foreign key type
-      //////////////////////////////////////////////////////////////////////////////////////////////
-      for (const loadedAssociation of loadedAssociations) {
-        dreams
-          .filter((dream: any) => {
-            return (
-              dream[association.foreignKeyTypeField()] === loadedAssociation['stiBaseClassOrOwnClassName'] &&
-              dream[association.foreignKey()] === association.primaryKeyValue(loadedAssociation)
-            )
-          })
-          .forEach((dream: any) => {
-            dream[association.as] = loadedAssociation
-          })
-      }
-      ///////////////////////////////////////////////////////////////////////////////////////////////////
-      // end: Associate each loaded association with each dream based on primary key and foreign key type
-      ///////////////////////////////////////////////////////////////////////////////////////////////////
-    }
-  }
-
-  /**
-   * @internal
-   *
-   * Applies a preload statement
-   */
-  private async applyOnePreload(
-    this: Query<DreamInstance, QueryTypeOpts>,
-    associationName: string,
-    dreams: Dream | Dream[],
-    onStatement: RelaxedPreloadOnStatement<any, any> = {}
-  ) {
-    if (!Array.isArray(dreams)) dreams = [dreams] as Dream[]
-
-    const dream = dreams.find(dream => dream['getAssociationMetadata'](associationName))!
-    if (!dream) return
-
-    const { name, alias } = extractAssociationMetadataFromAssociationName(associationName)
-
-    const association = dream['getAssociationMetadata'](name)
-    if (association === undefined) throw new UnexpectedUndefined()
-
-    const dreamClass = dream.constructor as typeof Dream
-    const dreamClassToHydrate = association.modelCB() as typeof Dream
-
-    if ((association.polymorphic && association.type === 'BelongsTo') || Array.isArray(dreamClassToHydrate))
-      return this.preloadPolymorphicBelongsTo(
-        association as BelongsToStatement<any, any, any, string>,
-        dreams
-      )
-
-    const dreamClassToHydrateColumns = [...dreamClassToHydrate.columns()]
-    const throughColumnsToHydrate: any[] = []
-
-    const columnsToPluck = dreamClassToHydrateColumns.map(column =>
-      this.namespaceColumn(column.toString(), alias)
-    ) as any[]
-
-    const asHasAssociation = association as
-      | HasManyStatement<any, any, any, any>
-      | HasOneStatement<any, any, any, any>
-
-    if (asHasAssociation.through && asHasAssociation.preloadThroughColumns) {
-      if (isObject(asHasAssociation.preloadThroughColumns)) {
-        const preloadMap = asHasAssociation.preloadThroughColumns as Record<string, string>
-        Object.keys(preloadMap).forEach(preloadThroughColumn => {
-          throughColumnsToHydrate.push(preloadMap[preloadThroughColumn])
-          columnsToPluck.push(this.namespaceColumn(preloadThroughColumn, asHasAssociation.through))
-        })
-      } else {
-        const preloadArray = asHasAssociation.preloadThroughColumns as string[]
-        preloadArray.forEach(preloadThroughColumn => {
-          throughColumnsToHydrate.push(preloadThroughColumn)
-          columnsToPluck.push(this.namespaceColumn(preloadThroughColumn, asHasAssociation.through))
-        })
-      }
-    }
-
-    columnsToPluck.push(this.namespaceColumn(dreamClass.primaryKey, dreamClass.table))
-
-    const baseClass = dreamClass['stiBaseClassOrOwnClass']['getAssociationMetadata'](associationName)
-      ? dreamClass['stiBaseClassOrOwnClass']
-      : dreamClass
-
-    const associationDataScope = this.dreamClassQueryWithScopeBypasses(baseClass, {
-      // In order to stay DRY, preloading leverages the association logic built into
-      // `joins` (by using `pluck`, which calls `joins`). However, baseClass may have
-      // default scopes that would preclude finding that instance. We remove all
-      // default scopes on baseClass, but not subsequent associations, so that the
-      // single query will be able to find each row corresponding to a Dream in `dreams`,
-      // regardless of default scopes on that Dream's class.
-      bypassAllDefaultScopesExceptOnAssociations: true,
-    }).where({
-      [dreamClass.primaryKey]: dreams.map(obj => obj.primaryKeyValue),
-    })
-
-    const hydrationData: any[][] = await associationDataScope
-      ._connection(this.connectionOverride)
-      .innerJoin(associationName, (onStatement || {}) as JoinAndStatements<any, any, any, any>)
-      .pluck(...columnsToPluck)
-
-    const preloadedDreamsAndWhatTheyPointTo: PreloadedDreamsAndWhatTheyPointTo[] = hydrationData.map(
-      pluckedData => {
-        const attributes = {} as any
-        dreamClassToHydrateColumns.forEach(
-          (columnName, index) =>
-            (attributes[protectAgainstPollutingAssignment(columnName)] = pluckedData[index])
-        )
-
-        const hydratedDream = sqlResultToDreamInstance(dreamClassToHydrate, attributes)
-
-        throughColumnsToHydrate.forEach(
-          (throughAssociationColumn, index) =>
-            ((hydratedDream as any).preloadedThroughColumns[throughAssociationColumn] =
-              pluckedData[dreamClassToHydrateColumns.length + index])
-        )
-
-        return {
-          dream: hydratedDream,
-          pointsToPrimaryKey: pluckedData.at(-1),
-        }
-      }
-    )
-
-    this.hydrateAssociation(dreams, association, preloadedDreamsAndWhatTheyPointTo)
-
-    return preloadedDreamsAndWhatTheyPointTo.map(obj => obj.dream)
-  }
-
-  /**
-   * @internal
-   *
-   * Used by loadBuider
-   */
-  private async hydratePreload(this: Query<DreamInstance, QueryTypeOpts>, dream: Dream) {
-    await this.applyPreload(this.preloadStatements as any, this.preloadOnStatements as any, dream)
-  }
-
-  /**
-   * @internal
-   *
-   * Applies a preload statement
-   */
-  private async applyPreload(
-    this: Query<DreamInstance, QueryTypeOpts>,
-    preloadStatement: RelaxedPreloadStatement,
-    preloadOnStatements: RelaxedPreloadOnStatement<any, any>,
-    dream: Dream | Dream[]
-  ) {
-    const keys = Object.keys(preloadStatement as any)
-
-    for (const key of keys) {
-      const nestedDreams = await this.applyOnePreload(
-        key,
-        dream,
-        this.applyablePreloadOnStatements(preloadOnStatements[key] as RelaxedPreloadOnStatement<any, any>)
-      )
-
-      if (nestedDreams) {
-        await this.applyPreload((preloadStatement as any)[key], preloadOnStatements[key] as any, nestedDreams)
-      }
-    }
-  }
-
-  /**
-   * @internal
-   *
-   * retrieves on statements that can be applied to a preload
-   */
-  private applyablePreloadOnStatements(
-    preloadOnStatements: RelaxedPreloadOnStatement<any, any> | undefined
-  ): RelaxedPreloadOnStatement<any, any> | undefined {
-    if (preloadOnStatements === undefined) return undefined
-
-    return Object.keys(preloadOnStatements).reduce(
-      (agg, key) => {
-        const value = preloadOnStatements[key]
-        if (value === undefined) throw new UnexpectedUndefined()
-        // filter out plain objects, but not ops and not and/andNot/andAny statements
-        // because plain objects are just the next level of nested preload
-        if (
-          key === 'and' ||
-          key === 'andNot' ||
-          key === 'andAny' ||
-          value === null ||
-          value.constructor !== Object
-        ) {
-          agg[key] = value
-        }
-
-        return agg
-      },
-      {} as RelaxedPreloadOnStatement<any, any>
-    )
-  }
-
-  private conditionallyApplyDefaultScopes() {
-    if (this.bypassAllDefaultScopes || this.bypassAllDefaultScopesExceptOnAssociations) return this
-
-    const thisScopes = this.dreamClass['scopes'].default
-    let query: Query<DreamInstance, QueryTypeOpts> = this
-    for (const scope of thisScopes) {
-      if (
-        !shouldBypassDefaultScope(scope.method, {
-          defaultScopesToBypass: [
-            ...this.defaultScopesToBypass,
-            ...this.defaultScopesToBypassExceptOnAssociations,
-          ],
-        })
-      ) {
-        query = (this.dreamClass as any)[scope.method](query)
-      }
-    }
-
-    return query
-  }
-
-  /**
-   * Each association in the chain is pushed onto `throughAssociations`
-   * and `applyOneJoin` is recursively called. The trick is that the
-   * through associations don't get written into the SQL; they
-   * locate the next association we need to build into the SQL,
-   * which is only run by the association that started the `through`
-   * chain. The final association at the end of the `through` chain _is_
-   * written into the SQL as a full association, but the modifications from
-   * the `through` association are only added when the recursion returns
-   * back to the association that kicked off the through associations.
-   */
-
-  private joinsBridgeThroughAssociations<
-    QueryType extends
-      | SelectQueryBuilder<any, any, any>
-      | UpdateQueryBuilder<any, any, any, any>
-      | DeleteQueryBuilder<any, any, any>,
-    Schema extends DreamInstance['schema'],
-  >({
-    query,
-    dreamClass,
-    association,
-    previousAssociationTableOrAlias,
-    throughAssociations,
-    joinType,
-  }: {
-    query: QueryType
-    dreamClass: typeof Dream
-    association: AssociationStatement
-    previousAssociationTableOrAlias: TableOrAssociationName<Schema>
-    throughAssociations: (HasOneStatement<any, any, any, any> | HasManyStatement<any, any, any, any>)[]
-    joinType: JoinTypes
-  }): {
-    query: QueryType
-    dreamClass: typeof Dream
-    association: AssociationStatement
-    throughClass?: typeof Dream | null
-    previousAssociationTableOrAlias: TableOrAssociationName<Schema>
-  } {
-    if (association.type === 'BelongsTo' || !association.through) {
-      return {
-        query,
-        dreamClass,
-        association,
-        previousAssociationTableOrAlias,
-      }
-    } else {
-      throughAssociations.push(association)
-
-      // We have entered joinsBridgeThroughAssociations with the
-      // CompositionAssetAudits HasOne User association, which
-      // is through compositionAsset
-      // We now apply the compositionAsset association (a BelongsTo)
-      // to the query
-      const { query: queryWithThroughAssociationApplied } = this.applyOneJoin({
-        query,
-        dreamClass,
-        previousAssociationTableOrAlias,
-        currentAssociationTableOrAlias: association.through as TableOrAssociationName<Schema>,
-        throughAssociations,
-        joinType,
-      })
-
-      // The through association has both a `through` and a `source`. The `source`
-      // is the association on the model that has now been joined. In our example,
-      // the `source` is the `user` association on the CompositionAsset model
-      const { newAssociation, throughAssociation, throughClass } = this.followThroughAssociation(
-        dreamClass,
-        association
-      )
-
-      if ((newAssociation as any).through) {
-        // This new association is itself a through association, so we recursively
-        // call joinsBridgeThroughAssociations
-        return this.joinsBridgeThroughAssociations({
-          query: queryWithThroughAssociationApplied,
-          dreamClass: throughClass,
-          association: newAssociation,
-          previousAssociationTableOrAlias: throughAssociation.as as TableOrAssociationName<Schema>,
-          throughAssociations,
-          joinType,
-        })
-      } else {
-        // This new association is not a through association, so
-        // this is the target association we were looking for
-        return {
-          query: queryWithThroughAssociationApplied,
-          dreamClass: association.modelCB(),
-          association: newAssociation,
-          throughClass,
-          previousAssociationTableOrAlias: association.through as TableOrAssociationName<Schema>,
-        }
-      }
-    }
-  }
-
-  private applyOneJoin<
-    QueryType extends
-      | SelectQueryBuilder<any, any, any>
-      | UpdateQueryBuilder<any, any, any, any>
-      | DeleteQueryBuilder<any, any, any>,
-    DB extends DreamInstance['DB'],
-    Schema extends DreamInstance['schema'],
-  >({
-    query,
-    dreamClass,
-    previousAssociationTableOrAlias,
-    currentAssociationTableOrAlias,
-    joinAndStatements = {},
-    throughAssociations = [],
-    joinType,
-  }: {
-    query: QueryType
-    dreamClass: typeof Dream
-    previousAssociationTableOrAlias: TableOrAssociationName<Schema>
-    currentAssociationTableOrAlias: TableOrAssociationName<Schema>
-    joinAndStatements?: RelaxedJoinAndStatement<any, any>
-    throughAssociations?: (HasOneStatement<any, any, any, any> | HasManyStatement<any, any, any, any>)[]
-
-    joinType: JoinTypes
-  }): {
-    query: QueryType
-    association: AssociationStatement
-    previousAssociationTableOrAlias: TableOrAssociationName<Schema>
-    currentAssociationTableOrAlias: TableOrAssociationName<Schema>
-  } {
-    const { name, alias } = extractAssociationMetadataFromAssociationName(currentAssociationTableOrAlias)
-    const joinAndStatement = joinAndStatements[currentAssociationTableOrAlias] as RelaxedJoinAndStatement<
-      DB,
-      Schema
-    >
-
-    previousAssociationTableOrAlias = extractAssociationMetadataFromAssociationName(
-      previousAssociationTableOrAlias
-    ).alias
-    currentAssociationTableOrAlias = alias
-
-    let association = dreamClass['getAssociationMetadata'](name)
-
-    if (!association) {
-      throw new JoinAttemptedOnMissingAssociation({
-        dreamClass,
-        associationName: currentAssociationTableOrAlias,
-      })
-    }
-
-    const results = this.joinsBridgeThroughAssociations({
-      query,
-      dreamClass,
-      association,
-      previousAssociationTableOrAlias,
-      throughAssociations,
-      joinType,
-    })
-
-    query = results.query
-    dreamClass = results.dreamClass
-
-    association = results.association
-    const timeToApplyThroughAssociations =
-      throughAssociations.length && throughAssociations[0]?.source === association.as
-
-    const originalPreviousAssociationTableOrAlias = previousAssociationTableOrAlias
-    previousAssociationTableOrAlias = results.previousAssociationTableOrAlias
-    const throughClass = results.throughClass
-
-    if (timeToApplyThroughAssociations) {
-      /**
-       * Each association in the chain is pushed onto `throughAssociations`
-       * and `applyOneJoin` is recursively called. The trick is that the
-       * through associations don't get written into the SQL; they
-       * locate the next association we need to build into the SQL,
-       * which is only run by the association that started the `through`
-       * chain (thus the
-       * `throughAssociations.length && throughAssociations[0].source === association.as`
-       * above). The final association at the end of the `through` chain _is_
-       * written into the SQL as a full association, but the modifications from
-       * the `through` association are only added when the recursion returns
-       * back to the association that kicked off the through associations.
-       */
-
-      throughAssociations.forEach(
-        (throughAssociation: HasOneStatement<any, any, any, any> | HasManyStatement<any, any, any, any>) => {
-          if (throughAssociation.type === 'HasMany') {
-            if ((query as SelectQueryBuilder<any, any, any>)?.distinctOn && throughAssociation.distinct) {
-              query = (query as SelectQueryBuilder<any, any, any>).distinctOn(
-                this.distinctColumnNameForAssociation({
-                  association: throughAssociation,
-                  tableNameOrAlias: currentAssociationTableOrAlias,
-                  foreignKey: throughAssociation.primaryKey(),
-                }) as string
-              ) as QueryType
-            }
-
-            if (throughAssociation.order) {
-              query = this.applyOrderStatementForAssociation({
-                query,
-                association: throughAssociation,
-                tableNameOrAlias: currentAssociationTableOrAlias,
-              })
-            }
-          }
-        }
-      )
-    }
-
-    if (association.type === 'BelongsTo') {
-      if (Array.isArray(association.modelCB()))
-        throw new CannotJoinPolymorphicBelongsToError({
-          dreamClass,
-          association,
-          innerJoinStatements: this.innerJoinStatements,
-          leftJoinStatements: this.leftJoinStatements,
-        })
-
-      const to = (association.modelCB() as typeof Dream).table
-      const joinTableExpression =
-        currentAssociationTableOrAlias === to
-          ? currentAssociationTableOrAlias
-          : `${to} as ${currentAssociationTableOrAlias}`
-
-      query = (query as any)[(joinType === 'inner' ? 'innerJoin' : 'leftJoin') as 'innerJoin'](
-        joinTableExpression,
-        (join: JoinBuilder<any, any>) => {
-          join = join.onRef(
-            this.namespaceColumn(association.foreignKey(), previousAssociationTableOrAlias),
-            '=',
-            this.namespaceColumn(association.primaryKey(), currentAssociationTableOrAlias)
-          )
-
-          if (timeToApplyThroughAssociations) {
-            throughAssociations.forEach(
-              (
-                throughAssociation: HasOneStatement<any, any, any, any> | HasManyStatement<any, any, any, any>
-              ) => {
-                join = this.applyAssociationAndStatementsToJoinStatement({
-                  join,
-                  association: throughAssociation,
-                  currentAssociationTableOrAlias,
-                  previousAssociationTableOrAlias: originalPreviousAssociationTableOrAlias,
-                  joinAndStatements,
-                })
-              }
-            )
-          }
-
-          join = this.conditionallyApplyDefaultScopesDependentOnAssociation({
-            join,
-            tableNameOrAlias: currentAssociationTableOrAlias,
-            association,
-          })
-
-          join = this.applyJoinAndStatement(join, joinAndStatement, currentAssociationTableOrAlias)
-
-          return join
-        }
-      )
-    } else {
-      const to = association.modelCB().table
-      const joinTableExpression =
-        currentAssociationTableOrAlias === to
-          ? currentAssociationTableOrAlias
-          : `${to} as ${currentAssociationTableOrAlias}`
-
-      query = (query as any)[(joinType === 'inner' ? 'innerJoin' : 'leftJoin') as 'innerJoin'](
-        joinTableExpression,
-        (join: JoinBuilder<any, any>) => {
-          join = join.onRef(
-            this.namespaceColumn(association.primaryKey(), previousAssociationTableOrAlias),
-            '=',
-            this.namespaceColumn(association.foreignKey(), currentAssociationTableOrAlias)
-          )
-
-          if (association.polymorphic) {
-            join = join.on((eb: ExpressionBuilder<any, any>) =>
-              this.whereStatementToExpressionWrapper(
-                eb,
-                this.aliasWhereStatement(
-                  {
-                    [association.foreignKeyTypeField()]: throughClass
-                      ? throughClass['stiBaseClassOrOwnClassName']
-                      : dreamClass['stiBaseClassOrOwnClassName'],
-                  } as any,
-                  currentAssociationTableOrAlias
-                )
-              )
-            )
-          }
-
-          if (timeToApplyThroughAssociations) {
-            throughAssociations.forEach(
-              (
-                throughAssociation: HasOneStatement<any, any, any, any> | HasManyStatement<any, any, any, any>
-              ) => {
-                join = this.applyAssociationAndStatementsToJoinStatement({
-                  join,
-                  association: throughAssociation,
-                  currentAssociationTableOrAlias,
-                  previousAssociationTableOrAlias: originalPreviousAssociationTableOrAlias,
-                  joinAndStatements,
-                })
-              }
-            )
-          }
-
-          join = this.applyAssociationAndStatementsToJoinStatement({
-            join,
-            association,
-            currentAssociationTableOrAlias,
-            previousAssociationTableOrAlias,
-            joinAndStatements,
-          })
-
-          join = this.conditionallyApplyDefaultScopesDependentOnAssociation({
-            join,
-            tableNameOrAlias: currentAssociationTableOrAlias,
-            association,
-          })
-
-          join = this.applyJoinAndStatement(join, joinAndStatement, currentAssociationTableOrAlias)
-
-          return join
-        }
-      )
-
-      if (association.type === 'HasMany') {
-        if (association.order) {
-          query = this.applyOrderStatementForAssociation({
-            query,
-            tableNameOrAlias: currentAssociationTableOrAlias,
-            association,
-          })
-        }
-
-        if ((query as SelectQueryBuilder<any, any, any>).distinctOn && association.distinct) {
-          query = (query as SelectQueryBuilder<any, any, any>).distinctOn(
-            this.distinctColumnNameForAssociation({
-              association,
-              tableNameOrAlias: currentAssociationTableOrAlias,
-              foreignKey: association.foreignKey(),
-            }) as string
-          ) as QueryType
-        }
-      }
-    }
-
-    return {
-      query,
-      association,
-      previousAssociationTableOrAlias,
-      currentAssociationTableOrAlias,
-    }
-  }
-
-  private applyAssociationAndStatementsToJoinStatement({
-    join,
-    currentAssociationTableOrAlias,
-    previousAssociationTableOrAlias,
-    association,
-    joinAndStatements,
-  }: {
-    join: JoinBuilder<any, any>
-    currentAssociationTableOrAlias: string
-    previousAssociationTableOrAlias: string
-    association: HasOneStatement<any, any, any, any> | HasManyStatement<any, any, any, any>
-    joinAndStatements: RelaxedJoinAndStatement<any, any>
-  }) {
-    if (association.and) {
-      this.throwUnlessAllRequiredWhereClausesProvided(
-        association,
-        currentAssociationTableOrAlias,
-        joinAndStatements
-      )
-
-      join = join.on((eb: ExpressionBuilder<any, any>) =>
-        this.whereStatementToExpressionWrapper(
-          eb,
-          this.aliasWhereStatement(
-            association.and as WhereStatement<any, any, any>,
-            currentAssociationTableOrAlias
-          ),
-          { disallowSimilarityOperator: false }
-        )
-      )
-    }
-
-    if (association.andNot) {
-      join = join.on((eb: ExpressionBuilder<any, any>) =>
-        this.whereStatementToExpressionWrapper(
-          eb,
-          this.aliasWhereStatement(
-            association.andNot as WhereStatement<any, any, any>,
-            currentAssociationTableOrAlias
-          ),
-          { negate: true }
-        )
-      )
-    }
-
-    if (association.andAny) {
-      join = join.on((eb: ExpressionBuilder<any, any>) =>
-        eb.or(
-          (association.andAny as WhereStatement<any, any, any>[]).map(whereAnyStatement =>
-            this.whereStatementToExpressionWrapper(
-              eb,
-              this.aliasWhereStatement(whereAnyStatement, currentAssociationTableOrAlias),
-              { disallowSimilarityOperator: false }
-            )
-          )
-        )
-      )
-    }
-
-    if (association.selfAnd) {
-      join = join.on((eb: ExpressionBuilder<any, any>) =>
-        this.whereStatementToExpressionWrapper(
-          eb,
-          this.rawifiedSelfOnClause({
-            associationAlias: association.as,
-            selfAlias: previousAssociationTableOrAlias,
-            selfAndClause: association.selfAnd as any,
-          })
-        )
-      )
-    }
-
-    if (association.selfAndNot) {
-      join = join.on((eb: ExpressionBuilder<any, any>) =>
-        this.whereStatementToExpressionWrapper(
-          eb,
-          this.rawifiedSelfOnClause({
-            associationAlias: association.as,
-            selfAlias: previousAssociationTableOrAlias,
-            selfAndClause: association.selfAndNot as any,
-          }),
-          { negate: true }
-        )
-      )
-    }
-
-    return join
-  }
-
-  private conditionallyApplyDefaultScopesDependentOnAssociation({
-    join,
-    tableNameOrAlias,
-    association,
-  }: {
-    join: JoinBuilder<any, any>
-    tableNameOrAlias: string
-    association: AssociationStatement
-  }) {
-    let scopesQuery = new Query<DreamInstance, QueryTypeOpts>(this.dreamInstance)
-    const associationClass = association.modelCB() as typeof Dream
-    const associationScopes = associationClass['scopes'].default
-
-    for (const scope of associationScopes) {
-      if (
-        !shouldBypassDefaultScope(scope.method, {
-          bypassAllDefaultScopes: this.bypassAllDefaultScopes,
-          defaultScopesToBypass: [...this.defaultScopesToBypass, ...(association.withoutDefaultScopes || [])],
-        })
-      ) {
-        const tempQuery = (associationClass as any)[scope.method](scopesQuery)
-        // The scope method on a Dream model should return a clone of the Query it receives
-        // (e.g. by returning `scope.where(...)`), but in case the function doesn't return,
-        // or returns the wrong thing, we check before overriding `scopesQuery` with what the
-        // method returned.
-        if (tempQuery && tempQuery.constructor === scopesQuery.constructor) scopesQuery = tempQuery
-      }
-    }
-
-    if (scopesQuery.whereStatements.length) {
-      join = join.on((eb: ExpressionBuilder<any, any>) =>
-        eb.and(
-          scopesQuery.whereStatements.flatMap(whereStatement =>
-            this.whereStatementToExpressionWrapper(
-              eb,
-              this.aliasWhereStatement(whereStatement, tableNameOrAlias),
-              { disallowSimilarityOperator: false }
-            )
-          )
-        )
-      )
-    }
-
-    return join
-  }
-
-  private distinctColumnNameForAssociation({
-    association,
-    tableNameOrAlias,
-    foreignKey,
-  }: {
-    association: any
-    tableNameOrAlias: string
-    foreignKey: string
-  }) {
-    if (!association.distinct) return null
-    if (association.distinct === true) return this.namespaceColumn(foreignKey, tableNameOrAlias)
-    return this.namespaceColumn(association.distinct, tableNameOrAlias)
-  }
-
-  private recursivelyJoin<
-    QueryType extends
-      | SelectQueryBuilder<any, any, any>
-      | UpdateQueryBuilder<any, any, any, any>
-      | DeleteQueryBuilder<any, any, any>,
-    Schema extends DreamInstance['schema'],
-  >({
-    query,
-    joinStatement,
-    joinAndStatements,
-    dreamClass,
-    previousAssociationTableOrAlias,
-    joinType,
-  }: {
-    query: QueryType
-    joinStatement: RelaxedJoinStatement
-    joinAndStatements: RelaxedJoinAndStatement<any, any>
-    dreamClass: typeof Dream
-    previousAssociationTableOrAlias: TableOrAssociationName<Schema>
-    joinType: 'inner' | 'left'
-  }): QueryType {
-    for (const currentAssociationTableOrAlias of Object.keys(joinStatement)) {
-      const results = this.applyOneJoin({
-        query,
-        dreamClass,
-        previousAssociationTableOrAlias,
-        currentAssociationTableOrAlias,
-        joinAndStatements,
-        joinType,
-      })
-
-      query = results.query
-      const association = results.association
-
-      query = this.recursivelyJoin({
-        query,
-        joinStatement: joinStatement[currentAssociationTableOrAlias] as any,
-        joinAndStatements: joinAndStatements[currentAssociationTableOrAlias] as any,
-        dreamClass: association.modelCB() as typeof Dream,
-        previousAssociationTableOrAlias: currentAssociationTableOrAlias,
-        joinType,
-      })
-    }
-
-    return query
-  }
-
-  private throwUnlessAllRequiredWhereClausesProvided(
-    association: HasOneStatement<any, any, any, any> | HasManyStatement<any, any, any, any>,
-    namespace: string,
-    joinAndStatements: RelaxedJoinAndStatement<any, any>
-  ) {
-    const andClause = association.and!
-    const columnsRequiringAndStatements = Object.keys(andClause).reduce((agg, column) => {
-      if (andClause[column] === DreamConst.required) agg.push(column)
-      return agg
-    }, [] as string[])
-
-    const missingRequiredWhereStatements = columnsRequiringAndStatements.filter(
-      column => (joinAndStatements[namespace] as any)?.and?.[column] === undefined
-    )
-
-    if (missingRequiredWhereStatements.length)
-      throw new MissingRequiredAssociationAndClause(association, missingRequiredWhereStatements[0])
-  }
-
-  private applyOrderStatementForAssociation<
-    QueryType extends
-      | SelectQueryBuilder<any, any, any>
-      | UpdateQueryBuilder<any, any, any, any>
-      | DeleteQueryBuilder<any, any, any>,
-  >({
-    query,
-    tableNameOrAlias,
-    association,
-  }: {
-    query: QueryType
-    tableNameOrAlias: string
-    association: HasManyStatement<any, any, any, any>
-  }): QueryType {
-    if (!(query as SelectQueryBuilder<any, any, any>).orderBy) return query
-    let selectQuery = query as SelectQueryBuilder<any, any, any>
-    const orderStatement = association.order
-
-    if (typeof orderStatement === 'string') {
-      selectQuery = selectQuery.orderBy(this.namespaceColumn(orderStatement, tableNameOrAlias), 'asc')
-    } else {
-      Object.keys(orderStatement as Record<string, OrderDir>).forEach(column => {
-        const direction = (orderStatement as any)[column] as OrderDir
-        selectQuery = selectQuery.orderBy(this.namespaceColumn(column, tableNameOrAlias), direction)
-      })
-    }
-
-    return selectQuery as QueryType
-  }
-
-  private inArrayWithNull_or_notInArrayWithoutNull_ExpressionBuilder(
-    eb: ExpressionBuilder<any, any>,
-    a: any,
-    b: KyselyComparisonOperatorExpression,
-    c: any[]
-  ): ExpressionWrapper<any, any, SqlBool> {
-    const isNullStatement = eb(a, 'is', null)
-    const compactedC = compact(c)
-    if (compactedC.length) return eb.or([eb(a, b, compactedC), isNullStatement])
-    // not in an empty array means match everything
-    if (b === 'not in') return sql<boolean>`TRUE` as unknown as ExpressionWrapper<any, any, SqlBool>
-    return isNullStatement
-  }
-
-  private inArrayWithoutNullExpressionBuilder(
-    eb: ExpressionBuilder<any, any>,
-    a: any,
-    b: KyselyComparisonOperatorExpression,
-    c: any[]
-  ): ExpressionWrapper<any, any, SqlBool> {
-    const isNotNullStatement = eb(a, 'is not', null)
-    const compactedC = compact(c)
-    if (compactedC.length) return eb.and([eb(a, 'in', compactedC), isNotNullStatement])
-    // in an empty array means match nothing
-    return sql<boolean>`FALSE` as unknown as ExpressionWrapper<any, any, SqlBool>
-  }
-
-  private notInArrayWithNullExpressionBuilder(
-    eb: ExpressionBuilder<any, any>,
-    a: any,
-    b: KyselyComparisonOperatorExpression,
-    c: any[]
-  ): ExpressionWrapper<any, any, SqlBool> {
-    const isNullStatement = eb(a, 'is not', null)
-    const compactedC = compact(c)
-
-    if (compactedC.length) return eb.and([eb(a, 'not in', compactedC), isNullStatement])
-    return isNullStatement
-  }
-
-  private whereStatementToExpressionWrapper(
-    eb: ExpressionBuilder<any, any>,
-    whereStatement: WhereStatement<any, any, any>,
-    {
-      negate = false,
-      disallowSimilarityOperator = true,
-    }: {
-      negate?: boolean
-      disallowSimilarityOperator?: boolean
-    } = {}
-  ): ExpressionWrapper<any, any, SqlBool> {
-    const clauses = compact(
-      Object.keys(whereStatement)
-        .filter(key => (whereStatement as any)[key] !== DreamConst.required)
-        .map(attr => {
-          const val = (whereStatement as any)[attr]
-
-          if (
-            (val as OpsStatement<any, any>)?.isOpsStatement &&
-            (val as OpsStatement<any, any>).shouldBypassWhereStatement
-          ) {
-            if (disallowSimilarityOperator) throw new Error('Similarity operator may not be used in whereAny')
-
-            // some ops statements are handled specifically in the select portion of the query,
-            // and should be ommited from the where clause directly
-            return
-          }
-
-          const { a, b, c, a2, b2, c2 } = this.dreamWhereStatementToExpressionBuilderParts(attr, val)
-
-          // postgres is unable to handle WHERE IN statements with blank arrays, such as in
-          // "WHERE id IN ()", meaning that:
-          // 1. If we receive a blank array during an IN comparison,
-          //    then we need to simply regurgitate a where statement which
-          //    guarantees no records.
-          // 2. If we receive a blank array during a NOT IN comparison,
-          //    then it is the same as the where statement not being present at all,
-          //    resulting in a noop on our end
-          //
-
-          if (Array.isArray(c)) {
-            if ((b === 'in' && c.includes(null)) || (b === 'not in' && !c.includes(null))) {
-              return this.inArrayWithNull_or_notInArrayWithoutNull_ExpressionBuilder(eb, a, b, c)
-            } else if (negate && b === 'in' && !c.includes(null)) {
-              return this.inArrayWithoutNullExpressionBuilder(eb, a, b, c)
-            } else if (b === 'not in' && c.includes(null)) {
-              return this.notInArrayWithNullExpressionBuilder(eb, a, b, c)
-            }
-
-            const compactedC = compact(c)
-
-            if (b === 'in' && compactedC.length === 0) {
-              // in an empty array means match nothing
-              return sql<boolean>`FALSE`
-            } else if (b === 'not in' && compactedC.length === 0) {
-              // not in an empty array means match everything
-              return sql<boolean>`TRUE`
-            } else {
-              return eb(a, b, compactedC)
-            }
-
-            //
-          } else if (b === '=' && c === null) {
-            return eb(a, 'is', null)
-
-            //
-          } else if (b === '!=' && c === null) {
-            return eb(a, 'is not', null)
-
-            //
-          } else if (b === '=' && negate) {
-            return eb.and([eb(a, '=', c), eb(a, 'is not', null)])
-
-            //
-          } else if (b === '!=' && c !== null) {
-            return eb.or([eb(a, '!=', c), eb(a, 'is', null)])
-
-            //
-          } else {
-            const expression = eb(a, b, c)
-            if (b2) return expression.and(eb(a2, b2, c2))
-            return expression
-          }
-        })
-    )
-
-    return negate ? eb.not(eb.parens(eb.and(clauses))) : eb.and(clauses)
-  }
-
-  private dreamWhereStatementToExpressionBuilderParts(attr: string, val: any) {
-    let a: any
-    let b: KyselyComparisonOperatorExpression
-    let c: any
-    let a2: any = null
-    let b2: KyselyComparisonOperatorExpression | null = null
-    let c2: any = null
-
-    if (val instanceof Function && val !== DreamConst.passthrough) {
-      val = val()
-    } else if (val === DreamConst.passthrough) {
-      const column = attr.split('.').at(-1)
-      if ((this.passthroughOnStatement as any)[column!] === undefined)
-        throw new MissingRequiredPassthroughForAssociationAndClause(column!)
-      val = (this.passthroughOnStatement as any)[column!]
-    }
-
-    if (val === null) {
-      a = attr
-      b = 'is'
-      c = val
-    } else if (['SelectQueryBuilder', 'SelectQueryBuilderImpl'].includes(val?.constructor?.name as string)) {
-      a = attr
-      b = 'in'
-      c = val
-    } else if (Array.isArray(val)) {
-      a = attr
-      b = 'in'
-      c = val.map(v =>
-        v instanceof DateTime || v instanceof CalendarDate
-          ? v.toSQL()
-          : typeof v === 'string'
-            ? normalizeUnicode(v)
-            : v
-      )
-    } else if (val instanceof CurriedOpsStatement) {
-      val = val.toOpsStatement(this.dreamClass, attr)
-      a = attr
-      b = val.operator
-      c = val.value
-    } else if (val instanceof OpsStatement) {
-      a = attr
-      b = val.operator as KyselyComparisonOperatorExpression
-      c = val.value
-    } else if (val instanceof Range) {
-      const rangeStart = val.begin
-      const rangeEnd = val.end
-      const excludeEnd = val.excludeEnd
-
-      if (rangeStart && rangeEnd) {
-        a = attr
-        b = '>='
-        c = rangeStart
-        a2 = attr
-        b2 = excludeEnd ? '<' : '<='
-        c2 = rangeEnd
-      } else if (rangeStart) {
-        a = attr
-        b = '>='
-        c = rangeStart
-      } else {
-        a = attr
-        b = excludeEnd ? '<' : '<='
-        c = rangeEnd
-      }
-    } else {
-      a = attr
-      b = '='
-      c = val
-    }
-
-    if (c instanceof DateTime || c instanceof CalendarDate) c = c.toSQL()
-    else if (typeof c === 'string') c = normalizeUnicode(c)
-
-    if (c2 instanceof DateTime || c2 instanceof CalendarDate) c2 = c2.toSQL()
-    else if (typeof c2 === 'string') c2 = normalizeUnicode(c2)
-
-    if (a && c === undefined) throw new CannotPassUndefinedAsAValueToAWhereClause(this.dreamClass, a)
-    if (a2 && c2 === undefined) throw new CannotPassUndefinedAsAValueToAWhereClause(this.dreamClass, a2)
-
-    return { a, b, c, a2, b2, c2 }
-  }
-
-  private applyJoinAndStatement<Schema extends DreamInstance['schema']>(
-    join: JoinBuilder<any, any>,
-    joinAndStatement: JoinAndStatements<any, any, any, any> | null,
-    rootTableOrAssociationAlias: TableOrAssociationName<Schema>
-  ) {
-    if (!joinAndStatement) return join
-
-    join = this._applyJoinAndStatements(join, joinAndStatement.and, rootTableOrAssociationAlias)
-    join = this._applyJoinAndStatements(join, joinAndStatement.andNot, rootTableOrAssociationAlias, {
-      negate: true,
-    })
-    join = this._applyJoinAndAnyStatements(join, joinAndStatement.andAny, rootTableOrAssociationAlias)
-
-    return join
-  }
-
-  private _applyJoinAndStatements<Schema extends DreamInstance['schema']>(
-    join: JoinBuilder<any, any>,
-    joinAndStatement: WhereStatement<any, any, any> | undefined,
-    rootTableOrAssociationAlias: TableOrAssociationName<Schema>,
-    {
-      negate = false,
-    }: {
-      negate?: boolean
-    } = {}
-  ) {
-    if (!joinAndStatement) return join
-
-    return join.on((eb: ExpressionBuilder<any, any>) =>
-      this.joinAndStatementToExpressionWrapper(joinAndStatement, rootTableOrAssociationAlias, eb, {
-        negate,
-        disallowSimilarityOperator: negate,
-      })
-    )
-  }
-
-  private _applyJoinAndAnyStatements<Schema extends DreamInstance['schema']>(
-    join: JoinBuilder<any, any>,
-    joinAndAnyStatement: WhereStatement<any, any, any>[] | undefined,
-    rootTableOrAssociationAlias: TableOrAssociationName<Schema>
-  ) {
-    if (!joinAndAnyStatement) return join
-    if (!joinAndAnyStatement.length) return join
-
-    return join.on((eb: ExpressionBuilder<any, any>) => {
-      return eb.or(
-        joinAndAnyStatement.map(joinAndStatement =>
-          this.joinAndStatementToExpressionWrapper(joinAndStatement, rootTableOrAssociationAlias, eb)
-        )
-      )
-    })
-  }
-
-  private joinAndStatementToExpressionWrapper<Schema extends DreamInstance['schema']>(
-    joinAndStatement: WhereStatement<any, any, any>,
-    rootTableOrAssociationAlias: TableOrAssociationName<Schema>,
-    eb: ExpressionBuilder<any, any>,
-    {
-      negate = false,
-      disallowSimilarityOperator = true,
-    }: {
-      negate?: boolean
-      disallowSimilarityOperator?: boolean
-    } = {}
-  ) {
-    return this.whereStatementToExpressionWrapper(
-      eb,
-
-      Object.keys(joinAndStatement).reduce((agg: any, key: any) => {
-        agg[this.namespaceColumn(key.toString(), rootTableOrAssociationAlias)] = joinAndStatement[key]
-        return agg
-      }, {}),
-
-      {
-        negate,
-        disallowSimilarityOperator,
-      }
-    )
-  }
-
-  private buildCommon<
-    QueryType extends
-      | SelectQueryBuilder<any, any, any>
-      | UpdateQueryBuilder<any, any, any, any>
-      | DeleteQueryBuilder<any, any, any>,
-  >(this: Query<DreamInstance, QueryTypeOpts>, kyselyQuery: QueryType): QueryType {
-    this.checkForQueryViolations()
-
-    const query = this.conditionallyApplyDefaultScopes()
-
-    if (!isEmpty(query.innerJoinStatements)) {
-      kyselyQuery = query.recursivelyJoin({
-        query: kyselyQuery,
-        joinStatement: query.innerJoinStatements,
-        joinAndStatements: query.innerJoinAndStatements,
-        dreamClass: query.dreamClass,
-        previousAssociationTableOrAlias: this.baseSqlAlias,
-        joinType: 'inner',
-      })
-    }
-
-    if (!isEmpty(query.leftJoinStatements)) {
-      kyselyQuery = query.recursivelyJoin({
-        query: kyselyQuery,
-        joinStatement: query.leftJoinStatements,
-        joinAndStatements: query.leftJoinAndStatements,
-        dreamClass: query.dreamClass,
-        previousAssociationTableOrAlias: this.baseSqlAlias,
-        joinType: 'left',
-      })
-    }
-
-    if (query.whereStatements.length || query.whereNotStatements.length || query.whereAnyStatements.length) {
-      kyselyQuery = (kyselyQuery as SelectQueryBuilder<any, any, any>).where(
-        (eb: ExpressionBuilder<any, any>) =>
-          eb.and([
-            ...this.aliasWhereStatements(query.whereStatements, query.baseSqlAlias).map(whereStatement =>
-              this.whereStatementToExpressionWrapper(eb, whereStatement, {
-                disallowSimilarityOperator: false,
-              })
-            ),
-
-            ...this.aliasWhereStatements(query.whereNotStatements, query.baseSqlAlias).map(
-              whereNotStatement =>
-                this.whereStatementToExpressionWrapper(eb, whereNotStatement, { negate: true })
-            ),
-
-            ...query.whereAnyStatements.map(whereAnyStatements =>
-              eb.or(
-                this.aliasWhereStatements(whereAnyStatements, query.baseSqlAlias).map(whereAnyStatement =>
-                  this.whereStatementToExpressionWrapper(eb, whereAnyStatement)
-                )
-              )
-            ),
-          ])
-      ) as QueryType
-    }
-
-    return kyselyQuery
-  }
-
-  private checkForQueryViolations(this: Query<DreamInstance, QueryTypeOpts>) {
-    const invalidWhereNotClauses = this.similarityStatementBuilder().whereNotStatementsWithSimilarityClauses()
-    if (invalidWhereNotClauses.length) {
-      const invalidWhereNotClause = invalidWhereNotClauses[0]
-
-      if (invalidWhereNotClause === undefined) throw new UnexpectedUndefined()
-
-      const { tableName, columnName, opsStatement } = invalidWhereNotClause
-      throw new CannotNegateSimilarityClause(tableName, columnName, opsStatement.value)
-    }
-  }
-
-  private aliasWhereStatements(whereStatements: Readonly<WhereStatement<any, any, any>[]>, alias: string) {
-    return whereStatements.map(whereStatement => this.aliasWhereStatement(whereStatement, alias))
-  }
-
-  private aliasWhereStatement(whereStatement: Readonly<WhereStatement<any, any, any>>, alias: string) {
-    return Object.keys(whereStatement).reduce((aliasedWhere, key) => {
-      aliasedWhere[this.namespaceColumn(key, alias)] = (whereStatement as any)[key]
-      return aliasedWhere
-    }, {} as any)
-  }
-
-  private rawifiedSelfOnClause<DB extends DreamInstance['DB'], Schema extends DreamInstance['schema']>({
-    associationAlias,
-    selfAlias,
-    selfAndClause,
-  }: {
-    associationAlias: string
-    selfAlias: string
-    selfAndClause: SelfOnStatement<any, DB, Schema, DreamInstance['table']>
-  }) {
-    const alphanumericUnderscoreRegexp = /[^a-zA-Z0-9_]/g
-    selfAlias = selfAlias.replace(alphanumericUnderscoreRegexp, '')
-
-    return Object.keys(selfAndClause).reduce((acc, key) => {
-      const selfColumn = selfAndClause[key]?.replace(alphanumericUnderscoreRegexp, '')
-      if (!selfColumn) return acc
-
-      acc[this.namespaceColumn(key, associationAlias)] = sql.raw(
-        `"${snakeify(selfAlias)}"."${snakeify(selfColumn)}"`
-      )
-      return acc
-    }, {} as any)
-  }
-
-  private buildDelete(this: Query<DreamInstance, QueryTypeOpts>): DeleteQueryBuilder<any, any, any> {
-    const kyselyQuery = this.dbFor('delete').deleteFrom(
-      this.baseSqlAlias as unknown as AliasedExpression<any, any>
-    )
-
-    const results = this.attachLimitAndOrderStatementsToNonSelectQuery(kyselyQuery as any)
-    return results.clone.buildCommon(results.kyselyQuery)
-  }
-
-  private buildSelect(
-    this: Query<DreamInstance, QueryTypeOpts>,
-    {
-      bypassSelectAll = false,
-      bypassOrder = false,
-      columns,
-    }: {
-      bypassSelectAll?: boolean
-      bypassOrder?: boolean
-      columns?: DreamColumnNames<DreamInstance>[]
-    } = {}
-  ): SelectQueryBuilder<any, any, any> {
-    let kyselyQuery: SelectQueryBuilder<any, any, any>
-
-    if (this.baseSelectQuery) {
-      kyselyQuery = (
-        this.connectionOverride
-          ? this.baseSelectQuery.connection(this.connectionOverride)
-          : this.baseSelectQuery
-      ).buildSelect({ bypassSelectAll: true })
-    } else {
-      const from =
-        this.baseSqlAlias === this.tableName ? this.tableName : `${this.tableName} as ${this.baseSqlAlias}`
-
-      kyselyQuery = this.dbFor('select').selectFrom(from)
-    }
-
-    if (this.distinctColumn) {
-      kyselyQuery = kyselyQuery.distinctOn(this.distinctColumn)
-    }
-
-    kyselyQuery = this.buildCommon(kyselyQuery)
-
-    kyselyQuery = this.conditionallyAttachSimilarityColumnsToSelect(kyselyQuery, {
-      bypassOrder: bypassOrder || !!this.distinctColumn,
-    }) as typeof kyselyQuery
-
-    if (this.orderStatements.length && !bypassOrder) {
-      this.orderStatements.forEach(orderStatement => {
-        kyselyQuery = kyselyQuery.orderBy(
-          this.namespaceColumn(orderStatement.column),
-          orderByDirection(orderStatement.direction)
-        )
-      })
-    }
-
-    if (this.limitStatement) kyselyQuery = kyselyQuery.limit(this.limitStatement)
-    if (this.offsetStatement) kyselyQuery = kyselyQuery.offset(this.offsetStatement)
-
-    if (columns) {
-      kyselyQuery = kyselyQuery.select(
-        this.columnsWithRequiredLoadColumns(columns).map(column => this.namespaceColumn(column))
-      )
-    } else if (!bypassSelectAll) {
-      kyselyQuery = kyselyQuery.selectAll(this.baseSqlAlias)
-    }
-
-    // even though we manually bypass explicit order statements above,
-    // associations can contain their own ordering systems. If we do not
-    // escape all orders, we can mistakenly allow an order clause to sneak in.
-    if (bypassOrder) kyselyQuery = kyselyQuery.clearOrderBy()
-
-    return kyselyQuery
-  }
-
-  private columnsWithRequiredLoadColumns(columns: string[]) {
-    return uniq(
-      compact([this.dreamClass.primaryKey, this.dreamClass['isSTIBase'] ? 'type' : null, ...columns])
-    )
-  }
-
-  private buildUpdate<DB extends DreamInstance['DB']>(
-    attributes: Updateable<DreamInstance['table']>
-  ): UpdateQueryBuilder<DB, any, any, object> {
-    let kyselyQuery = this.dbFor('update')
-      .updateTable(this.tableName as DreamInstance['table'])
-      .set(attributes as any)
-
-    kyselyQuery = this.conditionallyAttachSimilarityColumnsToUpdate(kyselyQuery)
-
-    const results = this.attachLimitAndOrderStatementsToNonSelectQuery(kyselyQuery as any)
-    return results.clone.buildCommon(results.kyselyQuery)
-  }
-
-  private attachLimitAndOrderStatementsToNonSelectQuery<
-    T extends Query<DreamInstance, QueryTypeOpts>,
-    QueryType extends UpdateQueryBuilder<any, any, any, any> | DeleteQueryBuilder<any, any, any>,
-  >(this: T, kyselyQuery: QueryType): { kyselyQuery: QueryType; clone: T } {
-    if (this.limitStatement || this.orderStatements.length) {
-      kyselyQuery = (kyselyQuery as any).where((eb: ExpressionBuilder<any, any>) => {
-        const subquery = this.nestedSelect(this.dreamInstance.primaryKey)
-
-        return eb(this.dreamInstance.primaryKey as any, 'in', subquery)
-      }) as typeof kyselyQuery
-
-      return {
-        kyselyQuery,
-        clone: this.clone<T>({ where: null, whereNot: null, order: null, limit: null }),
-      }
-    }
-
-    return { kyselyQuery, clone: this }
-  }
-
-  private get hasSimilarityClauses() {
-    return (this as any).similarityStatementBuilder().hasSimilarityClauses
-  }
-
-  private similarityStatementBuilder(this: Query<DreamInstance, QueryTypeOpts>) {
-    return new SimilarityBuilder(this.dreamInstance, {
-      where: [...this.whereStatements],
-      whereNot: [...this.whereNotStatements],
-      joinAndStatements: this.innerJoinAndStatements,
-      transaction: this.dreamTransaction,
-      connection: this.connectionOverride,
-    })
-  }
-
-  private conditionallyAttachSimilarityColumnsToSelect(
-    this: Query<DreamInstance, QueryTypeOpts>,
-    kyselyQuery: SelectQueryBuilder<DreamInstance['DB'], any, object>,
-    { bypassOrder = false }: { bypassOrder?: boolean } = {}
-  ) {
-    const similarityBuilder = this.similarityStatementBuilder()
-    if (similarityBuilder.hasSimilarityClauses) {
-      kyselyQuery = similarityBuilder.select(kyselyQuery, { bypassOrder })
-    }
-
-    return kyselyQuery
-  }
-
-  private conditionallyAttachSimilarityColumnsToUpdate(
-    this: Query<DreamInstance, QueryTypeOpts>,
-    kyselyQuery: UpdateQueryBuilder<DreamInstance['DB'], any, any, any>
-  ) {
-    const similarityBuilder = this.similarityStatementBuilder()
-    if (similarityBuilder.hasSimilarityClauses) {
-      kyselyQuery = similarityBuilder.update(kyselyQuery)
-    }
-    return kyselyQuery
+    return await this.dbDriverInstance().update(attributes)
   }
 
   private invertOrder() {
@@ -4061,13 +2008,4 @@ export interface QueryOpts<
   transaction?: DreamTransaction<Dream> | null | undefined
   connection?: DbConnectionType | undefined
   shouldReallyDestroy?: boolean | undefined
-}
-
-function getSourceAssociation(dream: Dream | typeof Dream | undefined, sourceName: string) {
-  if (!dream) return
-  if (!sourceName) return
-  return (
-    (dream as Dream)['getAssociationMetadata'](sourceName) ||
-    (dream as Dream)['getAssociationMetadata'](pluralize.singular(sourceName))
-  )
 }
