+++ src/dream/QueryDriver/Postgres.ts
@@ -1,0 +1,6 @@
+import Dream from '../../Dream.js'
+import KyselyQueryDriver from './Kysely.js'
+
+export default class PostgresQueryDriver<
+  DreamInstance extends Dream,
+> extends KyselyQueryDriver<DreamInstance> {}
