+++ src/dream/internal/saveDream.ts
@@ -1,11 +1,9 @@
-import _db from '../../db/index.js'
 import Dream from '../../Dream.js'
 import ValidationError from '../../errors/ValidationError.js'
 import { DateTime } from '../../helpers/DateTime.js'
-import namespaceColumn from '../../helpers/namespaceColumn.js'
 import sqlAttributes from '../../helpers/sqlAttributes.js'
 import DreamTransaction from '../DreamTransaction.js'
-import executeDatabaseQuery from './executeDatabaseQuery.js'
+import Query from '../Query.js'
 import runHooksFor from './runHooksFor.js'
 
 export default async function saveDream<DreamInstance extends Dream>(
@@ -13,8 +11,6 @@ export default async function saveDream<DreamInstance extends Dream>(
   txn: DreamTransaction<Dream> | null = null,
   { skipHooks = false }: { skipHooks?: boolean } = {}
 ): Promise<DreamInstance> {
-  const db = txn?.kyselyTransaction ?? _db('primary')
-
   const alreadyPersisted = dream.isPersisted
 
   if (!skipHooks) {
@@ -31,29 +27,19 @@ export default async function saveDream<DreamInstance extends Dream>(
 
   if (alreadyPersisted && !dream.isDirty) return dream
 
-  let query: any
-
   const now = DateTime.now()
   if (!alreadyPersisted && !(dream as any).createdAt && dream.columns().has('createdAt'))
     (dream as any).createdAt = now
   if (!(dream.dirtyAttributes() as any).updatedAt && dream.columns().has('updatedAt'))
     (dream as any).updatedAt = now
 
-  const sqlifiedAttributes = sqlAttributes(dream)
-
-  if (alreadyPersisted) {
-    query = db
-      .updateTable(dream.table)
-      .set(sqlifiedAttributes as any)
-      .where(namespaceColumn(dream.primaryKey, dream.table), '=', dream.primaryKeyValue)
-  } else {
-    query = db.insertInto(dream.table).values(sqlifiedAttributes as any)
-  }
+  const hasUnsavedData = !!Object.keys(sqlAttributes(dream)).length
 
   // BeforeSave/Update actions may clear all the data that we intended to save, leaving us with
   // an invalid update command. The Sortable decorator is an example of this.
-  if (!alreadyPersisted || Object.keys(sqlifiedAttributes).length) {
-    const data = await executeDatabaseQuery(query.returning([...dream.columns()]), 'executeTakeFirstOrThrow')
+  if (!alreadyPersisted || hasUnsavedData) {
+    const data = await Query.dbDriverClass().saveDream(dream, txn)
+
     dream['isPersisted'] = true
     dream.setAttributes(data)
   }
